Create ltwtrdr.py
Run this in Python to (try to) get a text-only version of a webpage.

Needs some refining, only using this to pull text from my blog for reading and revising and it gets the job done.  Does not filter out some code (e.g. isogram, stylesheets), and encounters some unicode-related errors.